Implement str32

+++ src/main/scala/spack/Main.scala
@@ -1,7 +1,7 @@
 package spack
 
 object Main extends App {
-  val result = parse(Array[Byte](0xc1))
+  val result = parse(Array[Byte](0xc1.toByte))
   println(result)
 }
